Add sh3_arc_file
This abstracts over the zlib API for reading from arc files.

Because it provides a method to read a std::string, the previously used
char* in the structs are replaced by std::string.
As a bonus, a memory leak is gone with this change.

sh3_arc_file currently does not have an API to return the number of
bytes read, which is why the Load() methods now simply return a bool
indicating success.