Overhaul sh3_arc_file error handling
A read_error struct is introduced, of which a reference to must be
passed to the sh3_arc_file.Read* functions.
The errors can now be stringified and are thus printed when encountering
one.

All sh3_arc_file.Read* functions now return how many bytes were read
instead of the error, which allows partial string reads.