Basic texture Load function
This should work (untested because I haven't copied *.arc to my /data/
dir, but it says it can't find tex.arc, so it is looking in the right
section). It currently only loads everything into pixelbuff, which is
VERY bad because there is a fair amount of metadata stored before the
pixel data begins.