Fix clang warnings
vfile::fpos became std::size_t. It also makes sense that we're able to
address the whole range of fsize, which is also std::size_t.

sh3_arc_vfile had two logic errors that were fixed. One is an invalid if
conditional that was unintendedly constant, and one if conditional
checked against the wrong variable.