Added way to calculate Projection Matrix
Added functionality in the OpenGL context to calculate the current
Projection Matrix according to the Camera's FOV and the current Window's
dimensions, as well as being able to set the near and far values of the
current view frustum (handy for the outdoor Silent Hill level of SH3 and SH2). This may eventually need to be moved, as I don't really see a point in passing the window to any game logic (perhaps to change the Window settings in game, though this is unlike the original).