Allow specifying buffer_object::Targets for VAO buffers
This is achieved by changing the meaning of what "Attributes" are.
Before they just referred to an enum with slots.
Now it's a struct containing the slots-enum as well as an array of
buffer_object::Targets, which is used to construct the buffers in the VAO
with the desired target type.