Singleton Template class
Basic templated singleton template class. This is used in instances
where we need one and ONLY one instance of a class, such as the engine
itself, or it's managers. Due to the nature of the original SH3 engine,
I'm going out on a limb (and especially from my research) and assuming we'regoing to need a fair few "global" variables such as these to satisfy the original data structures.

This implementation is threadsafe. The instance is a static local
variable. According to the C++ standard:

`§6.7 [stmt.dcl] p4 If control enters the declaration concurrently while the variable is being initialized, the concurrent execution shall wait for completion of the initialization.`

If is therefore, threadsafe.