Complete overhaul of build system
Build system has been overhauled, especially for systems that do NOT
have a package manager installed (i.e Windows). The `libs` folder
contains the most up to date versions of libraries we are using, which
are known to be WORKING. This way (at least for Windows builds), we can
control versioning of our dependencies much better instead of having the
user have to fetch them. The premake5 file also takes into account which
system the user is on, meaning we can link linux versions of the
libraries, as WELL as windows versions.

Premake is also much less of a fuck around, and the simple `.lua` script
is very simple to modify and understand.