Shader system overhaul
Complete overhaul and simplification of the shader system. This version
only deals with Vertex and Fragment shaders in the final Shader Program;
there is no need or want for anything more. It is also now possible to
set and get Uniform values, in which the function is deduced during
compile time, which shaves off a bit of time required to perform a
uniform lookup.