Initial Engine class implementation
This is the starts to what will be the main engine class. By main
engine, we mean that is has control over the game, deals with 'scene'/state
switching etc and loading and cahing assets for later use.

There is also a very simple game loop included, though this is subject
to change. Hopefully we should be able to get menu up and running really
soon (within the week).