Updated Engine class
Engine class now properly operns a window handle. The main game loop
also now pumps the SDL2 even queue, of which the result is then passed
into `InputHandler()` in the currently running game state.

Move away from constructor initialisation

Window and Video subsystem critical structures have been moved away from
constructor initilisation so that they can be initialised within the
respective `init()` function found in `engine.cpp`. This makes it a lot
easier to control the sequence in which initialise certain things (I was
getting a lot of VERY strange segfaults, doing this seems to have solved
it).

Some people would probably argue against this, however, as it's not
necessarily ""C++'y"". This makes it easier to read, which C++ isn't
exactly known for...

Moved bitmap structures to texture

Unified `msbmp` and `CTexture`, as it makes more sense having the one
structure instead of two. At this point, it should be possible to have
our first actual 'scene' (of type `CGameState`, being the intro state).

Push to get engine up and running

Too many changes here. This was a bit of a scramble to get something
actually drawn on the screen. Still a long way off from having an actual
game, but this is a very promising start! :)

Updated Engine class

Engine class now properly operns a window handle. The main game loop
also now pumps the SDL2 even queue, of which the result is then passed
into `InputHandler()` in the currently running game state.

Move away from constructor initialisation

Window and Video subsystem critical structures have been moved away from
constructor initilisation so that they can be initialised within the
respective `init()` function found in `engine.cpp`. This makes it a lot
easier to control the sequence in which initialise certain things (I was
getting a lot of VERY strange segfaults, doing this seems to have solved
it).

Some people would probably argue against this, however, as it's not
necessarily ""C++'y"". This makes it easier to read, which C++ isn't
exactly known for...

Push to get engine up and running

Too many changes here. This was a bit of a scramble to get something
actually drawn on the screen. Still a long way off from having an actual
game, but this is a very promising start! :)